requirements.txt
Django==1.6.5
South==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-haystack==2.2.0
django-toolbelt==0.0.1
elasticsearch==1.2.0
gunicorn==19.1.0
psycopg2==2.5.3
pyelasticsearch==0.6.1
requests==2.3.0
simplejson==3.6.0
six==1.7.3
static3==0.5.1
urllib3==1.9
wsgiref==0.1.2
